refactor(src): refactor into fs, utils, and filters
