+++ src/filters/mod.rs
@@ -1,0 +1,7 @@
+mod file_type_filter;
+mod size_filter;
+mod time_filter;
+
+pub use file_type_filter::{FileTypeFilter, FileTypeFilterParser};
+pub use size_filter::{SizeFilter, SizeFilterParser};
+pub use time_filter::{TimeFilter, TimeFilterParser};
