+++ src/util/utils.rs
@@ -1,9 +1,7 @@
 use crate::dirent64;
-use crate::memchr_derivations::memrchr;
+use crate::util::memchr_derivations::memrchr;
 use core::ops::Deref;
 
-#[inline]
-#[cfg(any(target_os = "linux", target_os = "android"))]
 /*
   Wrapper for direct getdents syscalls
 
@@ -23,11 +21,13 @@ use core::ops::Deref;
  - 0: End of directory
  - Negative: Error code (check errno)
 */
+#[inline]
+#[cfg(any(target_os = "linux", target_os = "android"))]
 pub unsafe fn getdents<T>(fd: i32, buffer_ptr: *mut T, buffer_size: usize) -> libc::c_long
 where
     T: crate::ValueType, //i8/u8
 {
-    // SAFETY:Syscall has no other implicit safety requirements beyond pointer validity
+    // SAFETY: Syscall has no other implicit safety requirements beyond pointer validity
     unsafe { libc::syscall(libc::SYS_getdents64, fd, buffer_ptr, buffer_size) }
 }
 
@@ -189,6 +189,30 @@ My Cat Diavolo is cute.
 
 */
 //cargo-asm --lib fdf::utils::dirent_const_time_strlen (put to inline(never) to display)
+
+/**
+ Returns the length of a `dirent64's d_name` string in constant time using
+ SWAR (SIMD within a register) bit tricks.
+
+ This function avoids branching and SIMD instructions, achieving O(1) time
+by reading the final 8 bytes of the structure and applying bit-masking
+ operations to locate the null terminator.
+
+# Safety
+ The caller must ensure:
+ `dirent` is a valid, non-null pointer to a `libc::dirent64`.
+
+# Performance
+This is one of the hottest paths when scanning directories. By eliminating
+ branches and unnecessary memory reads, it improves efficiency compared with
+ conventional approaches.
+
+
+ # References
+ - [Stanford Bit Twiddling Hacks find 0 byte ](http://www.icodeguru.com/Embedded/Hacker%27s-Delight/043.htm)
+ - [find crate `dirent.rs`](https://github.com/Soveu/find/blob/master/src/dirent.rs)
+
+*/
 #[inline]
 #[cfg(any(
     target_os = "linux",
