+++ src/walk/mod.rs
@@ -1,0 +1,7 @@
+mod finder;
+mod finder_builder;
+mod types;
+
+pub use finder::Finder;
+pub use finder_builder::FinderBuilder;
+pub(crate) use types::{DirEntryFilter, FilterType};
