+++ src/walk/types.rs
@@ -1,0 +1,6 @@
+use crate::{SearchConfig, fs::DirEntry};
+
+/// Filter function type for directory entries,
+pub type FilterType = fn(&SearchConfig, &DirEntry, Option<DirEntryFilter>) -> bool;
+/// Generic filter function type for directory entries
+pub type DirEntryFilter = fn(&DirEntry) -> bool;
